+++ tests/examples/defer_cap/main.sh
@@ -38,6 +38,6 @@ grep "stepup.core.exceptions.InputNotFoundError: Dynamic inputs are not availabl
 
 # Wait for background processes, if any.
 set +e; wait -fn $PID; RETURNCODE=$?; set -e
-[[ "${RETURNCODE}" -eq $((RETURN_CODE_FAILED | RETURN_CODE_ONHOLD)) ]] || exit 1
+[[ "${RETURNCODE}" -eq $((RETURN_CODE_FAILED | RETURN_CODE_DRAINED)) ]] || exit 1
 
 [[ ! -f never.txt ]] || exit 1