+++ tests/examples/input_check_two/main.sh
@@ -18,4 +18,4 @@ wait
 [[ -f f1.txt ]] || exit 1
 [[ -f f2.txt ]] || exit 1
 [[ ! -f f3.txt ]] || exit 1
-grep "The scheduler has been put on hold" .stepup/fail.log
+grep "The scheduler is draining" .stepup/fail.log